RDST Utils to other utils, removing float32 test, fixing mutable parameters for RDST
